refactor: move wire bindings out of host
