+++ crates/dcc-mcp-host/Cargo.toml
@@ -13,9 +13,6 @@ tokio = { workspace = true }
 parking_lot = { workspace = true }
 thiserror = { workspace = true }
 tracing = { workspace = true }
-serde_json = { workspace = true }
-dcc-mcp-wire = { workspace = true }
-dcc-mcp-pybridge = { workspace = true, optional = true }
 workspace-hack = { version = "0.1", path = "../workspace-hack" }
 
 pyo3 = { workspace = true, optional = true }
@@ -31,5 +28,5 @@ pyo3 = { workspace = true, features = ["auto-initialize"] }
 
 [features]
 default = []
-python-bindings = ["pyo3", "dep:dcc-mcp-pybridge", "dcc-mcp-pybridge/python-bindings"]
+python-bindings = ["pyo3"]
 stub-gen = ["python-bindings", "dep:pyo3-stub-gen", "dep:pyo3-stub-gen-derive"]
