+++ python/dcc_mcp_core/_exports.py
@@ -585,8 +585,8 @@ _LAZY: dict[str, str] = {
     "sidecar_readiness_status": "dcc_mcp_core.install_lifecycle",
     "stop_runtime_entries": "dcc_mcp_core.install_lifecycle",
     "wait_for_sidecar_ready": "dcc_mcp_core.install_lifecycle",
-    "normalize_tool_arguments": "dcc_mcp_core.host",
-    "normalize_tool_meta": "dcc_mcp_core.host",
+    "normalize_tool_arguments": "dcc_mcp_core.wire",
+    "normalize_tool_meta": "dcc_mcp_core.wire",
     "SidecarActionDispatcher": "dcc_mcp_core.sidecar",
     "SidecarDispatchRequest": "dcc_mcp_core.sidecar",
     "ERROR_DISPATCH_FAILED": "dcc_mcp_core.sidecar",
