+++ python/dcc_mcp_core/host/__init__.py
@@ -38,13 +38,15 @@ except ImportError:
 from dcc_mcp_core.host._adapter import HostAdapter
 from dcc_mcp_core.host._adapter import TickableDispatcher
 from dcc_mcp_core.host._standalone import StandaloneHost
-from dcc_mcp_core.host._wire import normalize_tool_arguments
-from dcc_mcp_core.host._wire import normalize_tool_meta
 from dcc_mcp_core.host.qt_dispatcher import QtCommandServer
 from dcc_mcp_core.host.qt_dispatcher import ServerHandle
 from dcc_mcp_core.host.qt_dispatcher import current_server
 from dcc_mcp_core.host.qt_dispatcher import start_qt_server
 from dcc_mcp_core.host.qt_dispatcher import stop_qt_server
+
+# Backward-compatible aliases; canonical ownership lives in dcc_mcp_core.wire.
+from dcc_mcp_core.wire import normalize_tool_arguments
+from dcc_mcp_core.wire import normalize_tool_meta
 
 __all__ = [
     "BlockingDispatcher",
