+++ tests/test_host_wire_python.py
@@ -3,8 +3,8 @@ from __future__ import annotations
 import pytest
 
 import dcc_mcp_core
-from dcc_mcp_core.host import normalize_tool_arguments
-from dcc_mcp_core.host import normalize_tool_meta
+from dcc_mcp_core.wire import normalize_tool_arguments
+from dcc_mcp_core.wire import normalize_tool_meta
 
 
 def test_normalize_tool_arguments_accepts_python_dict() -> None:
@@ -51,6 +51,14 @@ def test_normalize_tool_meta_rejects_non_object_shapes() -> None:
         normalize_tool_meta("42")
 
 
-def test_top_level_lazy_exports_host_wire_helpers() -> None:
+def test_top_level_lazy_exports_wire_helpers() -> None:
     assert dcc_mcp_core.normalize_tool_arguments('{"x": 1}') == {"x": 1}
     assert dcc_mcp_core.normalize_tool_meta(None) is None
+
+
+def test_host_keeps_backward_compatible_wire_aliases() -> None:
+    from dcc_mcp_core.host import normalize_tool_arguments as host_normalize_arguments
+    from dcc_mcp_core.host import normalize_tool_meta as host_normalize_meta
+
+    assert host_normalize_arguments('{"legacy": true}') == {"legacy": True}
+    assert host_normalize_meta(None) is None
